requirements.txt
alabaster==0.7.10
astroid==2.1.0
autopep8==1.4.3
Babel==2.5.3
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
Django==2.0.5
django-crispy-forms==1.9.2
django-recaptcha==2.0.5
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.5
docutils==0.14
idna==2.6
imagesize==1.0.0
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
logilab-common==1.4.2
MarkupSafe
mccabe==0.6.1
numpy
packaging==17.1
pandas
psycopg2
pycodestyle==2.4.0
Pygments==2.2.0
pylint==2.2.2
pylint-django==2.0.4
pylint-plugin-utils==0.4
pyparsing==2.2.0
python-dateutil==2.7.5
pytz==2018.4
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.4
sphinx-rtd-theme==0.3.1
sphinxcontrib-websupport==1.0.1
sqlparse==0.3.0
typed-ast
urllib3==1.22
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.1.5
dronekit